requirements.txt
opencv-python>=4.9.0.80
diffusers==0.31.0
transformers>=4.49.0
tokenizers>=0.20.3
accelerate>=1.1.1
tqdm
datasets
imageio
easydict
ftfy
dashscope
imageio-ffmpeg
# numpy==1.24.4
wandb[media]
omegaconf
einops
av==13.1.0
opencv-python
git+https://github.com/openai/CLIP.git
open_clip_torch
starlette
pycocotools
lmdb
matplotlib
sentencepiece
pydantic==2.10.6
scikit-image
huggingface_hub[cli]
dominate
nvidia-pyindex
nvidia-tensorrt
pycuda
onnx
onnxruntime
onnxscript
onnxconverter_common
flask
flask-socketio
torchao
peft